ft(view-result): User can view result of elections - The user can see the outcome of elections [Delivers #163673556]
