+++ migration/candidate.js
@@ -8,9 +8,9 @@ const db = new Pool();
 const createTableCandidate = () => {
   const query = `CREATE TABLE IF NOT EXISTS candidate(
           id SERIAL NOT NULL,
-          party INT NOT NULL REFERENCES party(id),
-          office INT NOT NULL REFERENCES office(id),
-          candidate INT NOT NULL REFERENCES users(id),
+          party INT NOT NULL REFERENCES party(id) ON DELETE CASCADE,
+          office INT NOT NULL REFERENCES office(id) ON DELETE CASCADE,
+          candidate INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
           PRIMARY KEY (office, candidate)
     )`;
 
@@ -23,6 +23,26 @@ const createTableCandidate = () => {
     });
 };
 
+const createTableVote = () => {
+  const query = `CREATE TABLE IF NOT EXISTS vote(
+            id SERIAL NOT NULL,
+            createdOn DATE NOT NULL DEFAULT CURRENT_DATE,
+            createdBy INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+            office INT NOT NULL REFERENCES office(id) ON DELETE CASCADE,
+            candidate INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+            PRIMARY KEY (office, createdBy)
+      )`;
+
+  db.query(query)
+    .then((results) => {
+      console.log(results);
+    })
+    .catch((error) => {
+      console.log(error);
+    });
+};
+
 createTableCandidate();
+createTableVote();
 
 db.end();
