+++ routes/userRouter.js
@@ -8,6 +8,7 @@ import createUserQuery from '../model/userQueries/createUser';
 import loginUserQuery from '../model/userQueries/loginUser';
 import candidateQuery from '../model/userQueries/createCandidate';
 import voteQuery from '../model/userQueries/createVote';
+import viewResultQuery from '../model/userQueries/viewResults';
 
 const userRouter = express.Router();
 
@@ -15,5 +16,6 @@ userRouter.post('/auth/signup', trim.body, userMiddleware.createUserCheck, creat
 userRouter.post('/auth/login', trim.body, userMiddleware.loginUserCheck, loginUserQuery);
 userRouter.post('/office/:id/register', trim.body, userMiddleware.candidateCheck, tokenAuth.tokenCheck, candidateQuery);
 userRouter.post('/votes/', trim.body, userMiddleware.voteCheck, tokenAuth.tokenCheck, voteQuery);
+userRouter.post('/office/:id/result ', tokenAuth.tokenCheck, viewResultQuery);
 
 export default userRouter;
